fix floating-point precision warning

+++ src/noa/core/signal/CTF.hpp
@@ -222,7 +222,7 @@ namespace noa::signal {
             const auto lambda = relativistic_electron_wavelength(voltage) * 1e10; // angstroms
             constexpr f64 PI = Constant<f64>::PI;
             m_k1 = static_cast<value_type>(PI * lambda);
-            m_k2 = static_cast<value_type>(PI * 0.5 * m_cs_angstroms * lambda * lambda * lambda);
+            m_k2 = static_cast<value_type>(PI * 0.5 * static_cast<f64>(m_cs_angstroms) * lambda * lambda * lambda);
         }
 
         constexpr void set_amplitude_fraction_() {
@@ -514,7 +514,7 @@ namespace noa::signal {
             const auto lambda = relativistic_electron_wavelength(voltage) * 1e10; // angstroms
             constexpr f64 PI = Constant<f64>::PI;
             m_k1 = static_cast<value_type>(PI * lambda);
-            m_k2 = static_cast<value_type>(PI * 0.5 * m_cs_angstroms * lambda * lambda * lambda);
+            m_k2 = static_cast<value_type>(PI * 0.5 * static_cast<f64>(m_cs_angstroms) * lambda * lambda * lambda);
         }
 
         constexpr void set_amplitude_fraction_() {
